requirements.txt
fastapi
uvicorn
streamlit
tensorflow
opencv-python
numpy
Pillow
tifffile
scikit-learn
